+++ frontend/src/components/DocumentListItem.js
@@ -1,9 +1,12 @@
 import React, { Component } from 'react'
+import Dropzone from 'react-dropzone'
+import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import DocumentPreview from './DocumentPreview'
 import { Button, Modal, ModalFooter } from 'reactstrap'
-import Upload from './Upload'
-import { downloadDocument } from '../utils/ApiWrapper'
+import { downloadDocument, uploadDocument, getAllDocuments } from '../utils/ApiWrapper'
+import { updateDocuments } from '../redux/modules/user'
+import { beginLoading, endLoading } from '../redux/modules/auth'
 import uploadImg from '../media/greyUpload.png'
 import downloadImg from '../media/downloadGrey.png'
 import visit from '../media/visit.png'
@@ -12,22 +15,42 @@ const mapStateToProps = state => ({
   isPM: state.user.isPM
 })
 
+const mapDispatchToProps = dispatch => {
+  return bindActionCreators(
+    {
+      beginLoading,
+      endLoading,
+      updateDocuments
+    },
+    dispatch
+  )
+}
 class DocumentListItem extends Component {
   constructor(props) {
     super(props)
 
     this.state = {
       document: this.props.document,
-      modal: false
+      files: []
     }
 
     this.handleDownloadClick = this.handleDownloadClick.bind(this)
-    this.handleUploadClick = this.handleUploadClick.bind(this)
-    this.toggle = this.toggle.bind(this)
+    this.onDrop = this.onDrop.bind(this)
   }
 
-  toggle = () => {
-    this.setState({ modal: !this.state.modal })
+  async onDrop(files) {
+    this.setState({
+      files
+    })
+    this.props.beginLoading()
+    await uploadDocument(this.state.files[0], this.state.files[0].name, this.state.document._id)
+    const documents = await getAllDocuments()
+    if (documents) {
+      this.props.updateDocuments(documents)
+    } else {
+      this.props.updateDocuments([])
+    }
+    this.props.endLoading()
   }
 
   handleDownloadClick() {
@@ -35,30 +58,11 @@ class DocumentListItem extends Component {
     downloadDocument(document.fileID)
   }
 
-  handleUploadClick() {
-    // Upload click handling
-    this.setState({ modal: !this.state.modal })
-  }
-
   render() {
     const { isPM } = this.props
-    const path = require('path')
 
     return (
       <>
-        <Modal isOpen={this.state.modal} toggle={this.toggle}>
-          <Upload docID={this.props.document._id} />
-          <ModalFooter>
-            <Button
-              className="invalidSearchButton"
-              onClick={e => {
-                this.toggle()
-              }}
-            >
-              Return
-            </Button>
-          </ModalFooter>
-        </Modal>
         <tr className="hoverable">
           <td data-testid="docClass">{this.state.document.docClassName}</td>
           <td data-testid="fileName">
@@ -79,9 +83,16 @@ class DocumentListItem extends Component {
               </Button>
             )}
             {!isPM && (
-              <Button color="transparent" onClick={this.handleUploadClick}>
-                <img className="buttonimg" src={uploadImg} />
-              </Button>
+              <Dropzone onDrop={this.onDrop}>
+                {({ getRootProps, getInputProps }) => (
+                  <section>
+                    <div {...getRootProps()}>
+                      <input {...getInputProps()} />
+                      <img className="buttonimg" src={uploadImg} />
+                    </div>
+                  </section>
+                )}
+              </Dropzone>
             )}
           </td>
         </tr>
@@ -90,4 +101,7 @@ class DocumentListItem extends Component {
   }
 }
 
-export default connect(mapStateToProps)(DocumentListItem)
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(DocumentListItem)
