+++ frontend/src/components/index.js
@@ -3,7 +3,6 @@ import Dashboard from './Dashboard'
 import DocumentList from './DocumentList'
 import DocumentListItem from './DocumentListItem'
 import Register from './Register'
-import Upload from './Upload'
 import SelectDocumentsPage from './SelectDocuments'
 import BoxPreview from './DocumentPreview'
 import PMMainPage from './PMMainPage'
@@ -15,7 +14,6 @@ export {
   DocumentList,
   DocumentListItem,
   LogIn,
-  Upload,
   BoxPreview,
   Register,
   LoginPage,
